+++ plugin-api/build.gradle
@@ -34,6 +34,7 @@ dependencies {
   api project(':ethereum:permissioning:plugin-api')
   api project(':services:kvstore:plugin-api')
   api project(':ethereum:p2p:plugin-api')
+  api project(':ethereum:eth:plugin-api-txpool')
   api project(':datatypes')
   api 'org.apache.commons:commons-lang3'
   api 'io.consensys.tuweni:tuweni-bytes'
@@ -66,7 +67,8 @@ tasks.register('checkAPICompatibility', me.champeau.gradle.japicmp.JapicmpTask) 
     project(':ethereum:permissioning:plugin-api').tasks.named('jar'),
     project(':crypto:plugin-api').tasks.named('jar'),
     project(':services:kvstore:plugin-api').tasks.named('jar'),
-    project(':ethereum:p2p:plugin-api').tasks.named('jar'))
+    project(':ethereum:p2p:plugin-api').tasks.named('jar'),
+    project(':ethereum:eth:plugin-api-txpool').tasks.named('jar'))
   newClasspath.from(sourceSets.main.runtimeClasspath)
 
   packageIncludes = [
