--- plugin-api/src/main/java/org/hyperledger/besu/plugin/services/txvalidator/PluginTransactionPoolValidatorFactory.java
@@ -1,29 +1,0 @@
-/*
- * Copyright contributors to Hyperledger Besu.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.hyperledger.besu.plugin.services.txvalidator;
-
-import org.hyperledger.besu.plugin.Unstable;
-
-/** Interface for a factory that creates transaction validators for txpool usage */
-@Unstable
-public interface PluginTransactionPoolValidatorFactory {
-
-  /**
-   * Create a transaction validator for txpool usage
-   *
-   * @return the transaction validator
-   */
-  PluginTransactionPoolValidator createTransactionValidator();
-}
