+++ settings.gradle
@@ -43,6 +43,7 @@ include 'ethereum:api'
 include 'ethereum:blockcreation'
 include 'ethereum:core'
 include 'ethereum:eth'
+include 'ethereum:eth:plugin-api-txpool'
 include 'ethereum:evmtool'
 include 'ethereum:mock-p2p'
 include 'ethereum:p2p'
